+++ base/libgit2/types.jl
@@ -1,75 +1,5 @@
 # This file is a part of Julia. License is MIT: http://julialang.org/license
-
-"""
-    @kwdef typedef
-
-This is a helper macro that automatically defines a keyword-based constructor for the type
-declared in the expression `typedef`, which must be a `type` or `immutable`
-expression. The default argument is supplied by declaring fields of the form `field::T
-= default`. If no default is provided then the default is provided by the `_kwdef(T)`
-function.
-"""
-macro kwdef(expr)
-    expr = macroexpand(expr) # to expand @static
-    T = expr.args[2]
-    params_ex = Expr(:parameters)
-    call_ex = Expr(:call, T)
-    _kwdef!(expr.args[3], params_ex, call_ex)
-    quote
-        $expr
-        $(esc(Expr(:call, T, params_ex))) = $(esc(call_ex))
-    end
-end
-
-# @kwdef helper function
-# mutates arguments inplace
-function _kwdef!(blk, params_ex, call_ex)
-    for i in eachindex(blk.args)
-        ei = blk.args[i]
-        isa(ei, Expr) || continue
-        if ei.head == :(=)
-            # val::Typ = defexpr
-            dec = ei.args[1] # val::Typ
-            def = ei.args[2] # defexpr
-            push!(params_ex.args, Expr(:kw, dec, def))
-            push!(call_ex.args, dec.args[1])
-            blk.args[i] = dec
-        elseif ei.head == :(::)
-            dec = ei # val::Typ
-            def = :(kwdef_val($(ei.args[2])))
-            push!(params_ex.args, Expr(:kw, dec, def))
-            push!(call_ex.args, dec.args[1])
-        elseif ei.head == :block
-            # can arise with use of @static inside type decl
-            _kwdef!(ei, params_ex, call_ex)
-        end
-    end
-    blk
-end
-
-
-
-"""
-    kwdef_val(T)
-
-The default value for use with the `@kwdef` macro. Returns:
-
- - null pointer for pointer types (`Ptr{T}`, `Cstring`, `Cwstring`)
- - zero for integer types
- - no-argument constructor calls (e.g. `T()`) for all other types
-
-"""
-function kwdef_val end
-
-kwdef_val{T}(::Type{Ptr{T}}) = Ptr{T}(C_NULL)
-kwdef_val(::Type{Cstring}) = Cstring(C_NULL)
-kwdef_val(::Type{Cwstring}) = Cwstring(C_NULL)
-
-kwdef_val{T<:Integer}(::Type{T}) = zero(T)
-
-kwdef_val{T}(::Type{T}) = T()
-
-
+import Base.@kwdef
 import .Consts: GIT_SUBMODULE_IGNORE, GIT_MERGE_FILE_FAVOR, GIT_MERGE_FILE
 
 const OID_RAWSZ = 20
